refactor: make accelerate executor module installable
Restructure the former `hypha-accelerate-driver`, add the `uv` build metadata, and seed the package so we can build a wheel and install it from that artifact.  Rename the module to `hypha-accelerate-executor` and move it to `executors/accelerate` to match the general naming convention.
