+++ .github/workflows/pr.yaml
@@ -47,7 +47,7 @@ jobs:
     name: Validate Python
     strategy:
       matrix:
-        projects: ["drivers/hypha-accelerate-driver", "drivers/aim-driver"]
+        projects: ["executors/accelerate", "drivers/aim-driver"]
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 #v4.2.2
       - id: validate-python-uv
