+++ .github/workflows/release.yaml
@@ -45,7 +45,7 @@ jobs:
     if: needs.get-next-version.outputs.new-release-published == 'true'
     strategy:
       matrix:
-        projects: ["drivers/hypha-accelerate-driver", "drivers/aim-driver"]
+        projects: ["executors/accelerate", "drivers/aim-driver"]
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 #v4.2.2
       - id: validate-python-uv
@@ -79,7 +79,7 @@ jobs:
       - get-next-version
     strategy:
       matrix:
-        projects: ["drivers/hypha-accelerate-driver"]
+        projects: ["executors/accelerate"]
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 #v4.2.2
       - id: build-python-uv
